need to fix itoa

+++ ft_strncmp.c
@@ -12,7 +12,7 @@
 
 #include "libft.h"
 
-int		ft_strncmp(const char *s1, const char *s2, unsigned int n)
+int		ft_strncmp(const char *s1, const char *s2, size_t n)
 {
 	while (n && *s1 && (*s1 == *s2))
 	{
